requirements.txt
aiohttp
aiosignal
aiosqlite
altair
annotated-types
anthropic
anyio
appnope
attrs
decorator
faiss-cpu
fastapi
filetype
httpx
ipykernel
jsonpatch
jsonpointer
jsonschema
langchain
langchain-community
langchain-core
langchain-experimental
langchain-google-genai
langgraph
mypy-extensions
nltk
numpy
openai
pandas
pydeck
pygments
pyjson5
python-dotenv
scikit-learn
sentence-transformers
six
streamlit
threadpoolctl
torch
tqdm
typing-extensions
uvicorn[standard]
